requirements.txt
fastapi~=0.97.0
uvicorn[standard]~=0.22.0
langchain~=0.0.0
beautifulsoup4~=4.12.0
openai~=0.27.0
pinecone-client~=2.2.0
httpx~=0.24.0
SQLAlchemy~=2.0.0
alembic~=1.11.0
python-dotenv~=1.0.0
